% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_epi_aux.R
\name{betas_sd_mean}
\alias{betas_sd_mean}
\title{Computes beta values, standard deviation and mean
to plot the epimutation}
\usage{
betas_sd_mean(gr)
}
\arguments{
\item{gr}{a GRanges object obtained from
\link[epimutacions]{create_GRanges_class} function.}
}
\value{
The function returns a list
containing the melted beta values,
the population mean and 1, 1.5, and 2 standard deviations
from the mean of the distribution.
}
\description{
Computes the beta values, population mean and
 1, 1.5, and 2 standard deviations from the mean of the distribution
 necessary to plot the epimutations.
}
