% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_epi_aux.R
\name{cols_names}
\alias{cols_names}
\title{Sets common column names in a data frame}
\usage{
cols_names(x, cpg_ids_col = FALSE)
}
\arguments{
\item{x}{a data frame containing the genomic ranges or
a DMR (a row of the results of \link[epimutacions]{epimutations} or
\link[epimutacions]{epimutations_one_leave_out} function).}

\item{cpg_ids_col}{a boolean, if TRUE the input data frame contains
the CpGs names column.}
}
\value{
The function returns a data frame containing the column names
to carry out the analysis without any error.
}
\description{
Sets common column names in a given data frame
containing the CpGs genomic ranges or a DMR (result of
\link[epimutacions]{epimutations} or
\link[epimutacions]{epimutations_one_leave_out} function).
}
