% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_epi_aux.R
\name{create_GRanges_class}
\alias{create_GRanges_class}
\title{Generates a GRanges object}
\usage{
create_GRanges_class(methy, cpg_ids)
}
\arguments{
\item{methy}{a \code{GenomicRatioSet} object containing the control and
case samples used in \link[epimutacions]{epimutations} or
\link[epimutacions]{epimutations_one_leave_out} function.}

\item{cpg_ids}{a character string specifying the name of the
CpGs in the DMR of interest.}
}
\value{
The function returns a GRanges object containing the
beta values and the genomic ranges of the CpGs of interest.
}
\description{
This function makes a GRanges object
from a \code{GenomicRatioSet}.
}
