% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_epi_aux.R
\name{UCSC_annotation}
\alias{UCSC_annotation}
\title{UCSC gene annotations}
\usage{
UCSC_annotation(genome = "hg19")
}
\arguments{
\item{genome}{genome asambly. Can be set as:
\code{'hg38'}, \code{'hg19'} and \code{'hg18'}.}
}
\value{
The function returns gene
annotations for the specified genome assembly.
}
\description{
UCSC gene annotations for a given genome assembly.
}
