% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_epi_aux.R
\name{UCSC_regulation}
\alias{UCSC_regulation}
\title{UCSC annotation}
\usage{
UCSC_regulation(genome, chr, from, to)
}
\arguments{
\item{genome}{genome asambly. Can be set as:
\code{'hg38'}, \code{'hg19'} and \code{'hg18'}.}

\item{chr}{a character string containing
the sequence names to be analysed.}

\item{from, to}{scalar, specifying the range of genomic coordinates.
Note that \code{from} cannot be larger than \code{to}.}
}
\value{
\code{UCSC_regulation} returns
a list containing CpG Islands, H3K27Ac and H3K4Me3 tacks.
}
\description{
UCSC annotations for CpG Islands, H3K27Ac and H3K4Me3
 for a given genome assembly and genomic coordinates.
}
